extract test-utils to file
