+++ tests/no-internal-barrel-imports.js
@@ -10,35 +10,7 @@ const ESLintTester = require("eslint").RuleTester;
 const BentleyESLintPlugin = require("../dist");
 const NoInternalBarrelImportsESLintRule =
   BentleyESLintPlugin.rules["no-internal-barrel-imports"];
-
-/** @param {string[]} strings */
-function dedent(strings) {
-  const textAssumingNoInterpolations = strings[0];
-  const codeLines = textAssumingNoInterpolations.split("\n");
-  if (codeLines.length <= 1) return textAssumingNoInterpolations;
-  const leftPadding = codeLines[1].match(/[\t ]+/)[0];
-  return codeLines
-    .slice(1, -1)
-    .map((l) => l.substring(leftPadding.length))
-    .join("\n");
-}
-
-/** allow specifying `only` and `skip` properties for easier debugging */
-function supportSkippedAndOnlyInTests(obj) {
-  const hasOnly =
-    obj.valid.some((test) => Boolean(test.only)) ||
-    obj.invalid.some((test) => Boolean(test.only));
-  const keepTest = (test) => (hasOnly ? test.only : !test.skip);
-  const stripExtraTags = (test) => {
-    delete test.skip;
-    delete test.only;
-    return test;
-  };
-  return {
-    valid: obj.valid.filter(keepTest).map(stripExtraTags),
-    invalid: obj.invalid.filter(keepTest).map(stripExtraTags),
-  };
-}
+const { supportSkippedAndOnlyInTests, dedent } = require("./test-utils");
 
 const fixtureDir = path.join(
   __dirname,
